Refactoring for Styleguidist
