+++ src/api/Search.js
@@ -1,7 +1,6 @@
 // @flow
 import SimpleQueryRequest from './SimpleQueryRequest';
 import QueryResponse from './QueryResponse';
-import SavedSearch from './SavedSearch';
 import AuthUtils from '../util/AuthUtils';
 import FetchUtils from '../util/FetchUtils';
 import ObjectUtils from '../util/ObjectUtils';
@@ -88,8 +87,13 @@ export default class Search {
    * @param count         the number of documents to return (e.g. page size)
    * @param updateResults will be called when the search is complete with the results or an error
    */
-  simpleSearch(query: string, queryLanguage: 'simple' | 'advanced', offset: number, count: number,
-    updateResults: (response: QueryResponse | null, error: string | null) => void) {
+  simpleSearch(
+    query: string,
+    queryLanguage: 'simple' | 'advanced',
+    offset: number,
+    count: number,
+    updateResults: (response: QueryResponse | null, error: string | null) => void,
+  ) {
     const request = new SimpleQueryRequest();
     request.rows = count;
     request.query = query;
@@ -103,7 +107,8 @@ export default class Search {
     fieldName: string,
     fieldValues: Array<string>,
     onCompletion: () => void,
-    onError: (error: string) => void): Promise<any> {
+    onError: (error: string) => void,
+  ): Promise<any> {
     return new Promise((resolve, reject) => {
       // Get session
       const connectUri = `${this.baseUri}/rest/ingestApi/connect`;
@@ -214,112 +219,6 @@ export default class Search {
     });
   }
 
-  updateSavedSearches(ss: SavedSearch, mode?: string, callback: () => void): Promise<any> {
-    return new Promise((resolve, reject) => {
-      // Get session
-      const connectUri = `${this.baseUri}/rest/ingestApi/connect`;
-      fetch(connectUri, { credentials: 'include' })
-        .then((connectResult) => {
-          connectResult
-            .json()
-            .then((json) => {
-              const sessionId = json;
-              const updateUri = `${this.baseUri}/rest/ingestApi/feedDocuments/${sessionId}`;
-
-              const headers = new Headers({
-                Accept: 'application/json',
-                'Content-Type': 'application/json',
-              });
-
-              const loggedDateTime = new Date().toISOString();
-              const username = AuthUtils.getLoggedInUserId();
-              const jsonRequest1 =
-                '{ "fields" : { "title" : ' +
-                `[ "${ss.title}" ], "query_s" : [ "${ss.query}" ], "query_string_s" : [ "${ss.queryString}" ], ` +
-                `"username_s" : [ "${username}" ], "date" : [ "${loggedDateTime}" ], "table" : [ ".savedSearches" ] }, ` +
-                `"id" : "${ss.id}", "zone" : ".metadata", `;
-              const jsonRequest2 = mode ? `"mode" : "${mode}" }` : '}';
-              const jsonRequest = jsonRequest1 + jsonRequest2;
-              const body = JSON.stringify(jsonRequest);
-              const params = {
-                method: 'POST',
-                headers,
-                body,
-                credentials: 'include',
-              };
-
-              const updateFetchRequest = new Request(updateUri, params);
-              fetch(updateFetchRequest)
-                .then((updateResult: Response) => {
-                  if (updateResult.ok) {
-                    // Now need to refresh the update
-                    const refreshUri = `${this.baseUri}/rest/ingestApi/refresh/${sessionId}`;
-                    fetch(refreshUri, { credentials: 'include' })
-                      .then((refreshResult: Response) => {
-                        if (refreshResult.ok) {
-                          // Now need to close the session
-                          const disconnectUri = `${this.baseUri}/rest/ingestApi/disconnect/${sessionId}`;
-                          fetch(disconnectUri, { credentials: 'include' })
-                            .then((disconnectResult: Response) => {
-                              if (disconnectResult.ok) {
-                                callback();
-                                resolve();
-                              } else {
-                                // The request came back other than a 200-type response code
-                                disconnectResult
-                                  .text()
-                                  .then((msg) => {
-                                    reject(new Error(`Error disconnecting from the ingest API: ${msg}`));
-                                  })
-                                  .catch(() => {
-                                    reject(new Error(`Error disconnecting from the ingest API: ${disconnectResult.statusText}`));
-                                  });
-                              }
-                            })
-                            .catch((error) => {
-                              reject(new Error(`Failed to disconnect from the ingest API: ${error}`));
-                            });
-                        } else {
-                          // The request came back other than a 200-type response code
-                          refreshResult
-                            .text()
-                            .then((msg) => {
-                              reject(new Error(`Failed to refresh the update: ${msg}`));
-                            })
-                            .catch(() => {
-                              reject(new Error(`Failed to refresh the update: ${refreshResult.statusText}`));
-                            });
-                        }
-                      })
-                      .catch((error) => {
-                        reject(new Error(`Failed to refresh the update: ${error}`));
-                      });
-                  } else {
-                    // The request came back other than a 200-type response code
-                    updateResult
-                      .text()
-                      .then((msg) => {
-                        reject(new Error(`Failed to update the field: ${msg}`));
-                      })
-                      .catch((error) => {
-                        reject(new Error(`Failed to update the field: ${error}`));
-                      });
-                  }
-                })
-                .catch((error: any) => {
-                  // Catch network-type errors from the updating fetch() call
-                  reject(new Error(`Failed to update the field: ${error}`));
-                });
-            })
-            .catch((error) => {
-              reject(new Error(`Failed to connect to the ingest API: ${error}`));
-            });
-        })
-        .catch((error) => {
-          reject(new Error(`Failed to connect to the ingest API: ${error}`));
-        });
-    });
-  }
   addOrDeleteDocument(jsonRequest: JSON, callback: () => void): Promise<any> {
     return new Promise((resolve, reject) => {
       // Get session
